{{ cookiecutter.repo_name }}/pyproject.toml
[tool.poetry]
name = "{{ cookiecutter.project_name }}"
version = "0.0.0"
description = "{{ cookiecutter.project_description }}"
authors = ["{{ cookiecutter.author_name }} <{{ cookiecutter.author_email }}>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "{{ cookiecutter.src_package_name }}", from = "src" },
]
# https://pypi.org/classifiers/
classifiers = [
    "Private :: Not For Public Release"
]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.0"
djangorestframework = "^3.11"
django-filter = "^2.2.0"
drf-yasg = "^1.20.0"
drft = "^0.0.0-alpha.1"
psycopg2-binary = "^2.8.6"
whitenoise = "^5.2.0"
django-storages = {extras = ["boto3"], version = "^1.11.1"}
django-compressor = "^2.4"
django-environ = "^0.4.5"
aws-xray-sdk = "^2.6.0"
mixbag = "^0.0.0-alpha.1"


[tool.poetry.dev-dependencies]
pytest = "^6"
black = "^20.8b1"
flake8 = "^3.8.4"
mypy = "^0.790"
pytest-sugar = "^0.9.4"
pytest-mock = "^3.4.0"
pytest-responses = "^0.4.0"
pytest-picked = "^0.4.4"
pytest-instafail = "^0.4.2"
freezegun = "^1.0.0"
factory-boy = "^3.1.0"
codecov = "^2.1.11"
coverage = {extras = ["toml"], version = "^5.3"}
ipython = "^7.19.0"
pre-commit = "^2.9.3"
django-coverage-plugin = "^1.8.0"
pytest-django = "^4.1.0"
python-dotenv = "^0.15.0"
django-debug-toolbar = "^3.2"
django-extensions = "^3.1.0"
click = "^7.1.2"
django-test-migrations = "^1.1.0"
Sphinx = "^3.4.1"
sphinx-sitemap = "^2.2.0"
sphinx-pyreverse = "^0.0.13"
sphinx-autodoc-typehints = "^1.11.1"
sphinx-autobuild = "^2020.9.1"
recommonmark = "^0.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
tests = "scripts.tests:main"
server = "scripts.server:main"
startapp = "scripts.startapp:main"
manage = "scripts.manage:main"
docs-autobuild = "scripts.docs_autobuild:main"
docs-build = "scripts.docs_build:main"

[tool.coverage.run]
source = [
    "{{ cookiecutter.src_package_name }}",
]
omit = [
    "tests/*",
    "factories/*"
]

[tool.coverage.report]
omit = [
    "tests/*",
    "factories/*",
    "src/{{ cookiecutter.src_package_name }}/manage.py",
    "src/{{ cookiecutter.src_package_name }}/wsgi.py",
]
skip_empty = true
show_missing = true
fail_under = 100

[tool.black]
line-length = 79

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
addopts = "-s"
python_files = [
    "test_*.py",
    "tests.py",
    "*_tests.py"
]
testpaths = [
    "./tests",
]
